Make query parameters appear only once in URI
Otherwise duplicates of URI parameters can appear. This can lead to
bugs when, for example, somebody expects invoking `cursor(param)` on
a builder object to always set the cursor to the supplied `param`.
But if `cursor()` had already been called previously, the second
`param` would just be appended to the URI (`?cursor=123&cursor=345`).
Since Horizon only considers the first cursor parameter which appears
in the URI this can lead to strange behavior.