Add helper for fetching timebounds (#249)
Addresses #232.

(1) Add an axios interceptor that tracks the server's reported time from headers
(2) Add a function, `ServerSdk.Server.prototype.fetchTimebounds(seconds: Number)`, that returns a `timebounds` object for `TransactionBuilder` based on the server time. If no server time is recorded locally, or the recorded time is older than 5 minutes, it'll attempt to go to the network; if THAT fails, use the local time.