Remove server integration tests (#267)
These tests have failed several times recently for reasons not related to the library: once, when Friendbot stopped responding, and again when testnet reset. The functions called are tested with mocked functions in unit tests (and if they aren't, they can and should be). So it's not super helpful to have tests that fail non-deterministically.