Fix #391: Remove instance check for MessageEvent on stream error.
Using `MessageEvent` will throw an error if you are running the SDK on
Node (this API is not present in Node).

This was added in the TypeScript migration to match the expected
interface.

Checking the `error` instance shouldn't be a concern on the SDK but
the consumer, is up to the consumer what to do with an
error. Additionally we shouldn't be hiding errors just because they
don't match a given interface.