Change mention of CDN browser usage to CDNJS (#467)
Change the CDN mentioned and exampled that can be used in browsers to CDNJS instead of jsDelivr.

We mention and provide an example with jsDelivr at the top of the README, and then an example with CDNJS further buried in the README. The jsDelivr example doesn't seem to work from my own testing and I couldn't figure out how to make it work. It looks like it what we have hosted there might be tailored for use on a backend not in a browser. The CDNJS worked well for me and we use it in the second example further down so we may as well use it here.