Fix typescript definition for `TransactionCallBuilder#transaction`
A `TransactionCallBuilder` is expected to return a collection of
`Transactions` when calling the method `call`, however, the
`.transaction(hash)` method didn't match such expectation since it
will return a single transaction.

This isn't a problem for people using vanilla JS, but people using
TypeScript won't be able to compile their projects, since they are
expecting a single object however TypeScript will tell them that the
return value is a collection of object.