README: Update URL of Stellar logo image
### What
Update the URL for the Stellar logo image displayed at the top of the
README.

### Why
A new version of the stellar.org website was deployed and the image we
were using is no longer available there. In stellar/js-stellar-sdk#496
@tyvdh fixed the broken link. In stellar/.github#9 we added the logo to
a GitHub repository so that we could reference it from one location that
won't be affected by website changes that is easily accessible by any
dev. Other repositories (stellar/stellar-protocol#569,
stellar/stellar-core#2459, stellar/go#2363) are going to use the image
from that central location and we should update this README to use the
same one too.