Handle close SSE message properly (#567)
Before terminating an SSE connection, Horizon sends a "close" message on the connection. This PR adds a handler which is invoked whenever a "close" message is sent by Horizon. When the handler is invoked it closes the event source and recreates it. With this handler we no longer get any error messages when Horizon terminates an SSE connection.