feat(contract-client): match CLI's default timeout
Even 60 seconds, as it was updated to in
2bcbc6a4acfb67078794059802a479b0b0473c0f, might be too short if someone
is cautiously reviewing the transaction in Freighter, or if they need ot
unlock their Ledger and walk through its process.

Five minutes is the same as the CLI. Maybe it's a reasonable standard to
adopt?

This commit also updates utils.ts to use our eslint settings.