fix: performing 2nd swap not working
The test added here is copying the first few steps of the previous test,
"alice swaps bob 10 A for 1 B". For some reason, the second time you
simulate this transaction, it is calculating that no non-invoker signing
is needed! Haven't figured out why yet.

Note that this is an order-dependent test. The new one only functions as
expected because mocha runs the tests in the same order every time.