Adds `AssembledTransaction.toXDR|fromXDR` and `Client.txFromXDR` methods
Adds a toXDR and fromXDR for AssembledTransaction class. 
Also adds a txFromXDR to the Client class. 
Changes the swap test to use this flow.

Limitations: If you use the XDR multi-auth flow, you must re-simulate 
before the final `signAndSend` call. This is due to the fact that we can't 
serialize the XDR of the transaction envelope with the results of the 
simulation. If we want to do this, we would need to add an 
`AssembledTransaction` type to XDR in `stellar-base`.
---------

Co-authored-by: George <Shaptic@users.noreply.github.com>