Absolutely BASED documentation rendering fixes (#965)
* update README badges
* remove conflicting class tags from call builder classesmd
* make WebAuth show up as a namespace
* collect methods and classes into a SorobanRpc namespace for JSDoc
* create a Horizon JSDoc namespace and add the server class to it
* Fix all broken links to API dev docs
* add the Durability enum to the SorobanRPC JSDoc namespace
* create a util namespace and add validateTimebounds to it
* remove unnecessary code blocks from ContractSpec examples
* group some const declarations into more relevant namespaces/classes
* make a stellartoml JSDoc namespace
* make federation a module, include the typedefs
* mark most of the "call builder" classes as private in JSDoc

Since the comments clearly say "don't make this class yourself,"
it might make sense to mark them as `@private` so they don't show
up in the documentation.

The doc pages for those classes can still be visited, but they
don't show up in the sidebar to the left, unless you specifically
navigate to one of them. (Or, if you click a link for what the
return type on the `Horizon.Server.effects()` method, for example

* mark the Ok and Err rust type classes as private in JSDoc

I don't _think_ they're accessible by SDK users, so I think it
makes sense to not render them in the docs?

* change StellarToml from a JSDoc namespace to a module
* change webauth from a JSDoc namespace to a module
* mark the AccountResponse class as private in JSDoc
* add soroban tx timeout const to the SorobanRpc.Server module
* try out a new JSDoc theme
* make contract a JSDoc module

There was a weird `_default` global that came up from this. Not
sure why, but this seems to have done the trick without anything
weird going on after.

* mark contract/utils functions as private in JSDoc
* break the errors classes into multiple files for JSDoc rendering
* add JSDoc type definition to friendbot api

because more type definitions are more better, right?

* add default value to max toml file size in JSDoc
* clarify note about when the timebounds are added to client txs