chore: use `requestAirdrop` to fund e2e accounts
@willemneal [pointed out](https://github.com/stellar/stellar-docs/pull/1101#pullrequestreview-2461887053)
that it's a bit of a bummer to create accounts with a pure-SDK interface
just to revert to `fetch` and a hard-coded Friendbot URL in order to
fund that account.

This uses `rpc.Server#requestAirdrop` to do the funding.

I find the name of this method a bit unexpected! I wonder if the
language here has evolved since the method was first added.