Disable experimental module detection so that tests work:
Man what's the point of semantic versioning if you're just gonna enable random
settings (an EXPERIMENTAL FLAG, no less) by default in minor releases? See the
patch nodes of the latest minor release of Node v20[^1] and Node v22, both of
which are causing tests to fail:

> Module syntax detection (the --experimental-detect-module flag) is now enabled
> by default. Use --no-experimental-detect-module to disable it if needed.
>
> Syntax detection attempts to run ambiguous files as CommonJS, and if the
> module fails to parse as CommonJS due to ES module syntax, Node.js tries again
> and runs the file as an ES module.

Like bruh DID I ASK? No. Thus I am forced to add `"type": "module"` to my
package.json. But no, this is not possible, because unfortunately we still use a
mix of syntax and still need to support downstream users being able to `require`
the module :facepalm:

So I am instead forced to pass the flag disabling the experimental flag (???) to
mocha so that it doesn't introduce breaking behavior to tests.

[^1]: https://github.com/nodejs/node/releases/tag/v20.19.0