Disable disabling the experimental flag on Node 18:
Of course there's more to the story...

The option DOESN'T EXIST on Node 18, which is the earliest version
we support, meaning we can't USE the flag on that run of the test
matrix.

This means we need degenerate bash fuckery to ONLY enable the flag
on Node > 18 runs, so we can't use the mocha options like a normal
developer and must instead write some of the jankiest inline shell
scripting I've ever done to toggle it.