fix(client#MethodOptions): add overrideable params
This supercedes #1283, adding `publicKey`, `signTransaction`, and
`signAuthEntry` as optional parameters on `MethodOptions`. This is
actually the type of each method added to a Client ([[1]]), and needs to
allow overriding these settings.

This will allow us to fix the generated TS Bindings and specify
`MethodOptions` as the `options` type, rather than
`AssembledTransactionOptions`, which causes errors since these settings
(as well as `contractId`) are all **required**, but should not be
required for Client methods.

  [1]: https://github.com/theahaco/js-stellar-sdk/blob/f1f81f7870affb444c1d2a289c1a272f01410346/src/contract/client.ts#L27